ATLAS-5365: Add rest notification authorization and shared error mappers,and move Jersey exception mappers to server-common with rest web.xml updates for Java 17 and JSON error parity with webapp.
