+++ server-common/src/main/java/org/apache/atlas/server/common/errors/AllExceptionMapper.java
@@ -15,9 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.atlas.web.errors;
 
-import org.apache.atlas.exception.NotFoundException;
+package org.apache.atlas.server.common.errors;
+
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.core.Response;
@@ -26,16 +26,22 @@ import javax.ws.rs.ext.Provider;
 
 import java.util.concurrent.ThreadLocalRandom;
 
+/**
+ * Exception mapper for Jersey.
+ */
 @Provider
 @Component
-public class NotFoundExceptionMapper implements ExceptionMapper<NotFoundException> {
+public class AllExceptionMapper implements ExceptionMapper<Exception> {
     @Override
-    public Response toResponse(NotFoundException e) {
+    public Response toResponse(Exception exception) {
         final long id = ThreadLocalRandom.current().nextLong();
 
+        // Log the response and use the error codes from the Exception
+        ExceptionMapperUtil.logException(id, exception);
+
         return Response
-                .status(Response.Status.NOT_FOUND)
-                .entity(ExceptionMapperUtil.formatErrorMessage(id, e))
+                .serverError()
+                .entity(ExceptionMapperUtil.formatErrorMessage(id, exception))
                 .build();
     }
 }
