+++ server-common/src/main/java/org/apache/atlas/server/common/errors/ExceptionMapperUtil.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.atlas.web.errors;
+package org.apache.atlas.server.common.errors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,16 +28,16 @@ public class ExceptionMapperUtil {
     }
 
     @SuppressWarnings("UnusedParameters")
-    protected static String formatErrorMessage(long id, Exception exception) {
+    public static String formatErrorMessage(long id, Exception exception) {
         return String.format("There was an error processing your request. It has been logged (ID %016x).", id);
     }
 
-    protected static void logException(long id, Exception exception) {
+    public static void logException(long id, Exception exception) {
         LOGGER.error(formatLogMessage(id, exception), exception);
     }
 
     @SuppressWarnings("UnusedParameters")
-    protected static String formatLogMessage(long id, Throwable exception) {
+    public static String formatLogMessage(long id, Throwable exception) {
         return String.format("Error handling a request: %016x", id);
     }
 }
